:wrench: refactor and add popover tip on form field labels
