+++ src/components/Form/RequestForm/subcomponents/Radios.js
@@ -1,29 +1,14 @@
 import React from 'react';
 import { Field, ErrorMessage } from 'formik';
-import { Form, OverlayTrigger, Popover } from 'react-bootstrap';
+import { Form } from 'react-bootstrap';
+import LabelTip from './shared/LabelTip';
 
 const Radios = (props) => {
     const { label, name, options, required, description, tip, ...rest } = props;
 
     return (
         <Form.Group controlId={name}>
-            <Form.Label>
-                <OverlayTrigger
-                    placement='top'
-                    overlay={
-                        tip ? <Popover>
-                            <Popover.Content>
-                                {tip}
-                            </Popover.Content>
-                        </Popover> : () => <div></div>
-                    }
-                >
-                    <div>
-                        {label}
-                        {required && <span className='required'>*</span>}
-                    </div>
-                </OverlayTrigger>
-            </Form.Label>
+            <LabelTip label={label} tip={tip} required={required} />
 
             <Form.Text className='mb-3' muted>{description}</Form.Text>
